requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.4
django-crispy-forms==2.0
django-taggit==4.0.0
gunicorn==21.2.0
Markdown==3.4.4
packaging==23.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
